sakai-10.5

+++ samigo/samigo-app/src/java/org/sakaiproject/tool/assessment/ui/listener/author/RemovePublishedAssessmentListener.java
@@ -1,6 +1,6 @@
 /**********************************************************************************
- * $URL: https://source.sakaiproject.org/svn/sam/tags/sakai-10.4/samigo-app/src/java/org/sakaiproject/tool/assessment/ui/listener/author/RemovePublishedAssessmentListener.java $
- * $Id: RemovePublishedAssessmentListener.java 107319 2012-04-17 13:39:00Z david.horwitz@uct.ac.za $
+ * $URL: https://source.sakaiproject.org/svn/sam/tags/sakai-10.5/samigo-app/src/java/org/sakaiproject/tool/assessment/ui/listener/author/RemovePublishedAssessmentListener.java $
+ * $Id: RemovePublishedAssessmentListener.java 318753 2015-05-08 20:19:11Z ottenhoff@longsight.com $
  ***********************************************************************************
  *
  * Copyright (c) 2004, 2005, 2006, 2007, 2008 The Sakai Foundation
@@ -46,22 +46,21 @@ import org.sakaiproject.tool.assessment.integration.helper.ifc.GradebookServiceH
 import org.sakaiproject.tool.assessment.services.assessment.PublishedAssessmentService;
 import org.sakaiproject.tool.assessment.ui.bean.author.AuthorBean;
 import org.sakaiproject.tool.assessment.ui.bean.author.PublishedAssessmentBean;
+import org.sakaiproject.tool.assessment.ui.bean.authz.AuthorizationBean;
 import org.sakaiproject.tool.assessment.ui.listener.util.ContextUtil;
 /**
  * <p>Title: Samigo</p>
  * <p>Description: Sakai Assessment Manager</p>
  * @author Ed Smiley
- * @version $Id: RemovePublishedAssessmentListener.java 107319 2012-04-17 13:39:00Z david.horwitz@uct.ac.za $
+ * @version $Id: RemovePublishedAssessmentListener.java 318753 2015-05-08 20:19:11Z ottenhoff@longsight.com $
  */
 
 public class RemovePublishedAssessmentListener
     implements ActionListener
 {
   private static Log log = LogFactory.getLog(RemovePublishedAssessmentListener.class);
-  private static final GradebookServiceHelper gbsHelper =
-		IntegrationContextFactory.getInstance().getGradebookServiceHelper();
-  private static final boolean integrated =
-		IntegrationContextFactory.getInstance().isIntegrated();
+  private static final GradebookServiceHelper gbsHelper = IntegrationContextFactory.getInstance().getGradebookServiceHelper();
+  private static final boolean integrated = IntegrationContextFactory.getInstance().isIntegrated();
   private CalendarServiceHelper calendarService = IntegrationContextFactory.getInstance().getCalendarServiceHelper();
 
   public RemovePublishedAssessmentListener()
@@ -77,8 +76,13 @@ public class RemovePublishedAssessmentListener
       log.debug("assessmentId = " + assessmentId); 	    
       PublishedAssessmentService assessmentService = new PublishedAssessmentService();
       //get assessment to see if it has a calendar event
-      PublishedAssessmentFacade assessment = assessmentService.getPublishedAssessment(
-				assessmentId.toString());
+      PublishedAssessmentFacade assessment = assessmentService.getPublishedAssessment(assessmentId.toString());
+
+      AuthorizationBean authzBean = (AuthorizationBean) ContextUtil.lookupBean("authorization"); 
+      if (!authzBean.isUserAllowedToDeleteAssessment(assessmentId, assessment.getCreatedBy(), true)) {
+        throw new IllegalArgumentException("User does not have permission to delete assessmentId " + assessmentId);
+      }
+
       assessmentService.removeAssessment(assessmentId, "remove");
       removeFromGradebook(assessmentId);
